rag/pyproject.toml
[project]
name = "rag-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
